requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.0.4
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
inflection==0.5.1
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
mccabe==0.7.0
pep8-naming==0.13.2
Pillow==10.0.0
psycopg2-binary==2.9.6
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
referencing==0.30.0
rpds-py==0.9.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
